+++ drivers/gnss/gnss_nmea0183_match.c
@@ -4,6 +4,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
+#include <zephyr/drivers/gnss/gnss_publish.h>
 #include <zephyr/kernel.h>
 #include <zephyr/modem/chat.h>
 
@@ -11,7 +12,6 @@
 
 #include "gnss_nmea0183.h"
 #include "gnss_nmea0183_match.h"
-#include "gnss_publish.h"
 
 static bool gnss_nmea0183_match_timed_out(struct gnss_nmea0183_match_data *data)
 {
