+++ drivers/gnss/gnss_nmea_generic.c
@@ -7,13 +7,13 @@
  */
 
 #include <zephyr/drivers/gnss.h>
+#include <zephyr/drivers/gnss/gnss_publish.h>
 #include <zephyr/modem/chat.h>
 #include <zephyr/modem/backend/uart.h>
 #include <zephyr/kernel.h>
 #include <zephyr/drivers/gpio.h>
 #include <string.h>
 
-#include "gnss_publish.h"
 #include "gnss_nmea0183.h"
 #include "gnss_nmea0183_match.h"
 #include "gnss_parse.h"
