+++ include/zephyr/drivers/gnss/gnss_publish.h
